+++ src/nodecontainer.cpp
@@ -31,9 +31,10 @@ NodeReply crazytrace::NodeContainer::get_reply(const NodeRequest& request)
 
                 // Both variables undergo a value check during initialization so
                 // that neither is greater than 255. It is therefore safe to
-                // convert them into an int.
-                const int reply_hoplimit = reached_node->get_hoplimit() -
-                                           static_cast<int>(route.size()) + 1;
+                // convert them into an uint8_t.
+                const uint8_t reply_hoplimit = static_cast<uint8_t>(
+                    reached_node->get_hoplimit() -
+                    static_cast<uint8_t>(route.size()) + 1);
                 if (reply_hoplimit <= 0)
                     return NodeReply(NodeReplyType::NOREPLY);
 
@@ -78,13 +79,14 @@ NodeReply crazytrace::NodeContainer::get_reply(const NodeRequest& request)
             else
             {
                 /* hoplimit exceeded */
-                const int reached_node_number =
-                    static_cast<int>(route.size()) - request.get_hoplimit();
+                const unsigned int reached_node_number =
+                    static_cast<unsigned int>(route.size()) -
+                    request.get_hoplimit();
                 const std::shared_ptr<NodeInfo>& reached_node =
                     route.at(reached_node_number);
 
-                const int reply_hoplimit =
-                    reached_node->get_hoplimit() - hoplimit + 1;
+                const uint8_t reply_hoplimit = static_cast<uint8_t>(
+                    reached_node->get_hoplimit() - hoplimit + 1);
                 if (reply_hoplimit <= 0)
                     return NodeReply(NodeReplyType::NOREPLY);
